refactor: 검색어가 2글자 이하일 경우 경고 메시지 추가

+++ frontend/src/pages/project-list/ProjectSearchBar/ProjectSearchBar.tsx
@@ -7,6 +7,11 @@ function ProjectSearchBar() {
   const { refetch } = useProjectList();
 
   const handleSearchSubmit = (value: string) => {
+    if (value.length <= 1) {
+      alert("검색어는 2글자 이상 입력해주세요.");
+      return;
+    }
+
     if (value === "") {
       params.deleteAll({ replace: true });
       refetch();
